+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/internal/git/HasGit.kt
@@ -7,4 +7,4 @@ internal val Project.hasGit: Boolean
     get() = gitDir.exists()
 
 internal val Project.gitDir: File
-    get() = file("${rootProject.projectDir}/.git")
+    get() = file("$rootDir/.git")