+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/tasks/WriteSemverTask.kt
@@ -73,7 +73,7 @@ constructor(
 
             writeSemverTask.configure { task ->
                 task.tagPrefix.set(project.projectTagPrefix)
-                task.version.set(project.provider { project.version.toString() })
+                task.version.set(project.version.toString())
             }
 
             project.tasks.namedLazily<CreateSemverTagTask>(CreateSemverTagTask.TaskName) { task ->