+++ semver-project-gradle-plugin/testFixtures/kotlin/com/javiersc/semver/project/gradle/plugin/Sandbox.kt
@@ -41,7 +41,7 @@ internal fun createSandboxFile(prefix: String): File =
 internal fun File.generateInitialCommitAddVersionTagAndAddNewCommit(
     doBefore: Git.() -> Unit = {},
     doAfter: Git.() -> Unit = {},
-): List<com.javiersc.semver.project.gradle.plugin.Commit> {
+): List<Commit> {
     createGitIgnore()
     val git: Git = Git.init().setDirectory(this).call()
     doBefore(git)
@@ -60,32 +60,29 @@ internal fun File.generateInitialCommitAddVersionTagAndAddNewCommit(
     doAfter(git)
 
     return listOf(
-        com.javiersc.semver.project.gradle.plugin.Commit(
+        Commit(
             message = "Add new2",
             fullMessage = "Add new2",
             hash = "f099fed42808c387d38f8e34934d4ef2d6f3c2c5",
             timestampEpochSecond = Instant.now().epochSecond,
             tags = emptyList()
         ),
-        com.javiersc.semver.project.gradle.plugin.Commit(
+        Commit(
             message = "Add new",
             fullMessage = "Add new",
             hash = "0d9be4d12a836435c91fc38e0d64bcd3c14c89c8",
             timestampEpochSecond = Instant.now().epochSecond,
             tags = emptyList()
         ),
-        com.javiersc.semver.project.gradle.plugin.Commit(
+        Commit(
             message = "Initial commit",
             fullMessage = "Initial commit",
             hash = "e2592aafc259bd797e32f1ea4fe60cc7c0698f70",
             timestampEpochSecond = Instant.now().epochSecond,
             tags =
-            listOf(
-                com.javiersc.semver.project.gradle.plugin.Tag(
-                    name = "1.0.0",
-                    refName = "refs/tags/1.0.0"
-                ),
-            )
+                listOf(
+                    Tag(name = "1.0.0", refName = "refs/tags/1.0.0"),
+                )
         )
     )
 }